inbox-assistant-backend/requirements.txt
fastapi
uvicorn[standard]
pdfplumber
spacy
google-genai
python-multipart
pydantic